backend/requirements.txt
# ----------------------------
# Core Frameworks
# ----------------------------
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1

# ----------------------------
# Database
# ----------------------------
sqlite3-binary==0.0.1  # SQLite for local dev (Python built-in, optional)

# ----------------------------
# Image Processing & AI Modules
# ----------------------------
opencv-python==4.10.0.84
numpy==1.26.4
torch==2.3.0
torchvision==0.18.0
ultralytics==8.2.0     # YOLOv11
pytesseract==0.3.10    # OCR text detection
Pillow==10.3.0

# ----------------------------
# Location & Maps
# ----------------------------
geopy==2.4.1
requests==2.32.3

# ----------------------------
# Utility & Server Tools
# ----------------------------
python-dotenv==1.0.1   # For environment variables
gunicorn==22.0.0       # For production server
whitenoise==6.6.0      # For static files in Django

# ----------------------------
# Optional: For Testing & Logging
# ----------------------------
pytest==8.2.0
loguru==0.7.2

# ----------------------------
# Machine Learning (optional future use)
# ----------------------------
scikit-learn==1.5.0
pandas==2.2.2
